feat: added edge coverage
