+++ libafl_unicorn/Cargo.toml
@@ -12,7 +12,8 @@ edition = "2021"
 categories = ["development-tools::testing", "emulators", "embedded", "os", "no-std"]
 
 [dependencies]
-#libafl = { path = "../libafl", default-features = false, version = "0.8.2", features = ["std", "libafl_derive", "frida_cli"] }
-#libafl_targets = { path = "../libafl_targets", version = "0.8.2", features = ["std", "sancov_cmplog"] }
+libafl = { path = "../libafl", version = "0.8.2", default-features = false, features = ["std", "derive", "llmp_compression"] }
+libafl_targets = { path = "../libafl_targets", version = "0.8.2" }
+hashbrown =  { version = "0.12", features = ["serde", "ahash-compile-time-rng"] } # A faster hashmap, nostd compatible
 
 unicorn-engine = "2.0.1"
